refactor: 기존 HomeScreen을 home 폴더로 이동
